Version 2.8 (Jan 10, 2015)
This is a beta version capable of expanding WikiMedia templates.
Expanding templates requires a double pass on the dump, one for
collecting the templates and one for performing extraction. Processing
templates hence can take considerably longer: it is possible though to
save extracted templates to a file choosing option `--temaplates FILE`
in order to avoid repeating the scan for templates.